+++ cdp-agentkit-core/typescript/src/actions/cdp/social/farcaster/index.ts
@@ -4,7 +4,7 @@
 
 import { FarcasterAction, FarcasterActionSchemaAny } from "./farcaster_action";
 import { FarcasterAccountDetailsAction } from "./account_details";
-import { PublishCastAction } from "./publish_cast";
+import { FarcasterPostCastAction } from "./post_cast";
 
 /**
  * Retrieve an array of Farcaster action instances.
@@ -12,7 +12,7 @@ import { PublishCastAction } from "./publish_cast";
  * @returns {FarcasterAction<FarcasterActionSchemaAny>[]} An array of Farcaster action instances.
  */
 export function getAllFarcasterActions(): FarcasterAction<FarcasterActionSchemaAny>[] {
-  return [new FarcasterAccountDetailsAction(), new PublishCastAction()];
+  return [new FarcasterAccountDetailsAction(), new FarcasterPostCastAction()];
 }
 
 /**
@@ -24,8 +24,8 @@ export const FARCASTER_ACTIONS = getAllFarcasterActions();
  * All Farcaster action types.
  */
 export {
-  FarcasterAccountDetailsAction,
   FarcasterAction,
   FarcasterActionSchemaAny,
-  PublishCastAction,
+  FarcasterAccountDetailsAction,
+  FarcasterPostCastAction,
 };