+++ cdp-agentkit-core/typescript/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Added `farcaster_account_details` action.
-- Added `publish_cast` action.
+- Added `farcaster_post_cast` action.
 - Added `deploy_contract` action to deploy a contract using the Solidity compiler
 
 ## [0.0.13] - 2025-01-22